src/requirements.txt
streamlit
boto3
numpy
scipy
Pillow
requests
